requirements.txt
selenium
html2text
langchain_groq
langchain_core
pyyaml
unittest
